requirements-dev.txt
pandas
flake8    #for code formatting
joblib
notebook
python-box
pyyaml
types-pyyaml
pillow
torch 
torchvision 
torchaudio
mlflow
fastapi